+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - `coinsort` now supports wildecards (\*) in recfolders, the parameter must be protected
 by single quote to avoid bash expantion 
+- `SubjectEndEP` and `SessionEndEP` plugins entry points for finalisation of session and subjects
 
 ### Fixed
 - EEG naming schema discrepency with BIDS standard